pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "level1-ofi-qr-strategy"
version = "0.1.0"
description = "Execution-aware research scaffold for Level-I order-flow imbalance and quote-revision studies."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pandas>=2.0", "PyYAML>=6.0", "wrds>=3.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
